Modified demo completely, creating a simpler demo of a Quotes application.
The Projects application is more appropriate for a three day course or something as such.

The lecture still needs work, it is missing parts, and the flow needs to be considered. I need to refocus on giving a taste of a better more flexible development, rather than truly try to teach, it is unreasonable to teach such large and complex topics in so little time.